refactor: move the get grpc option from jina helper to runtime helper
